float-chat-ai/backend/requirements.txt
# Core FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Data processing
xarray>=2023.12.0
netCDF4>=1.6.4
pandas>=2.1.4
numpy>=1.21.0
scipy>=1.11.4

# Database
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.13.1

# AI/ML (optional - using fallbacks)
# openai>=1.6.1

# Data visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0

# HTTP and utilities
httpx>=0.25.2
python-multipart>=0.0.6
python-dotenv>=1.0.0

# CORS and middleware
fastapi-cors>=0.0.6

# Logging and monitoring
loguru>=0.7.2
